WORK/alloydb-query-bench-alloydb-only/requirements.txt

SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
PyYAML>=6.0.1
pandas>=2.2.0
numpy>=1.26.0
tabulate>=0.9.0
gspread>=6.0.0
google-auth>=2.29.0
